+++ packages/implementations/react/src/index.ts
@@ -2,6 +2,7 @@
 export * from "@rarui-react/box";
 export * from "@rarui-react/button";
 export * from "@rarui-react/checkbox";
+export * from "@rarui-react/chip";
 export * from "@rarui-react/icon";
 export * from "@rarui-react/link";
 export * from "@rarui-react/skeleton";