+++ packages/implementations/vuejs/src/atomic/Badge/src/Badge.vue
@@ -1,8 +1,5 @@
 <template>
-  <div
-    :class="badge.classnames.badge({ appearance, size, variant })"
-    v-bind="props"
-  >
+  <div :class="badge.classnames.badge({ appearance, size, variant })">
     <slot v-if="props.size != 'dot'" />
   </div>
 </template>