+++ packages/styles/src/packages/atomic/badge/rarui-badge.css.ts
@@ -1,6 +1,10 @@
 import { recipe } from "@vanilla-extract/recipes";
 import { varsThemeBase } from "../../../themes";
 
+/* -------------------------------------------------------------------------------------------------
+ * Style
+ * -----------------------------------------------------------------------------------------------*/
+
 const badge = recipe({
   base: {
     display: "flex",
@@ -60,16 +64,19 @@ const badge = recipe({
         padding: `0 ${varsThemeBase.spacing.xs}`,
         height: "3rem",
         fontSize: varsThemeBase.fontSize.button.l,
+        fontWeight: varsThemeBase.fontWeight.semiBold,
       },
       medium: {
         padding: `0 ${varsThemeBase.spacing["2xs"]}`,
         height: "2em",
         fontSize: varsThemeBase.fontSize.button.m,
+        fontWeight: varsThemeBase.fontWeight.medium,
       },
       small: {
         padding: `0 ${varsThemeBase.spacing["3xs"]}`,
         height: "1.5rem",
         fontSize: varsThemeBase.fontSize.button.s,
+        fontWeight: varsThemeBase.fontWeight.medium,
       },
       dot: {
         height: "0.5rem",