+++ packages/styles/src/packages/atomic/box/rarui-box.types.ts
@@ -1,0 +1,3 @@
+import { sprinkle } from "./rarui-box.css";
+
+export type BoxSprinkle = Parameters<typeof sprinkle>[0];