+++ packages/styles/src/packages/atomic/button/rarui-button.css.ts
@@ -3,6 +3,10 @@ import { recipe } from "@vanilla-extract/recipes";
 import { varsThemeBase } from "../../../themes";
 import { buttonVariants } from "./rarui-button.commons";
 
+/* -------------------------------------------------------------------------------------------------
+ * Style
+ * -----------------------------------------------------------------------------------------------*/
+
 const button = recipe({
   base: {
     display: "flex",
@@ -13,7 +17,6 @@ const button = recipe({
     cursor: "pointer",
     boxSizing: "border-box",
     position: "relative",
-    overflow: "hidden",
     gap: varsThemeBase.spacing["4xs"],
     fontWeight: varsThemeBase.fontWeight.semiBold,
     lineHeight: varsThemeBase.lineHeight.button.l,
@@ -104,7 +107,6 @@ const button = recipe({
     },
   },
   compoundVariants: [
-    // variant outlined
     {
       variants: {
         appearance: "brand",
@@ -282,7 +284,7 @@ const overlay = style({
   right: -1,
   pointerEvents: "none",
   opacity: 0,
-  borderRadius: varsThemeBase.shape.border.radius["2xs"],
+  borderRadius: "inherit",
 });
 
 globalStyle(`${button()}:hover ${overlay}`, {