+++ packages/styles/src/packages/atomic/chip/rarui-chip.types.ts
@@ -1,4 +1,5 @@
 import { RecipeVariants } from "@vanilla-extract/recipes";
-import { styles } from "./rarui-chip.css";
+import { styles, sprinkle } from "./rarui-chip.css";
 
 export type ChipVariants = RecipeVariants<typeof styles.chip>;
+export type ChipSprinkle = Parameters<typeof sprinkle>[0];