+++ packages/styles/src/packages/atomic/text/rarui-text.types.ts
@@ -1,0 +1,3 @@
+import { sprinkle } from "../box/rarui-box.css";
+
+export type TextSprinkle = Parameters<typeof sprinkle>[0];